requirements-prod.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.11.0
openai>=1.0.0
chromadb>=0.4.0
python-dotenv>=1.0.0
